Remove DrumRollCirclePiece, cleanup CirclePiece a bit.
