+++ osu.Game.Modes.Taiko/osu.Game.Modes.Taiko.csproj
@@ -64,7 +64,6 @@
     <Compile Include="Objects\Drawable\DrawableDrumRollTick.cs" />
     <Compile Include="Objects\Drawable\DrawableSwell.cs" />
     <Compile Include="Objects\Drawable\DrawableTaikoHitObject.cs" />
-    <Compile Include="Objects\Drawable\Pieces\DrumRollCirclePiece.cs" />
     <Compile Include="Objects\Drawable\Pieces\RimHitSymbolPiece.cs" />
     <Compile Include="Objects\Drawable\Pieces\StrongCirclePiece.cs" />
     <Compile Include="Objects\Drawable\Pieces\CirclePiece.cs" />
